fix(gjk): Exactly overlapping identical shapes caused degeneracy in GJK.
The cause was a dual one: the first search vector was the zero vector, causing the support points to be the zero vector also.
The second cause was that the lines between the support points passed through the origin, which cause the search vector to again become the zero vector.

Fixes #51.